Refactors part one
